set my observers in separate folder and gave users an  easy overview of their available credits
