--- src/js/api-functions/logout/logout.mjs
@@ -1,4 +1,0 @@
-export function logout() {
-  localStorage.clear();
-  window.location.href = "/index.html";
-}
