+++ src/scss/components/_header.scss
@@ -26,10 +26,19 @@ header {
   .logo_banner {
     width: 85%;
     display: grid;
-    grid-template-columns: 5fr 1fr 1fr;
+    grid-template-columns: 5fr 2fr 1fr;
     grid-auto-flow: row;
     justify-items: center;
     z-index: 3;
+
+    .users-credits {
+      background-color: map-get($map: variables.$colors, $key: offwhite);
+      padding: 10px;
+      text-align: center;
+      font-weight: 600;
+      border-radius: 25px;
+      border: 2px solid map-get($map: variables.$colors, $key: deepblue);
+    }
   }
 
   #hamburger_menu {
